LtrQueryBuilder should support the rewrite step
It was missed but is needed by some queries because they'll fail
if they are not rewritten.
This changes my mind concerning the model I'm not sure now
that it's reasonable to consider that FeatureSet.toQueries()
will accept a QueryShardContext and a map of params, it's likely
that templating will happen before.