Add 2 new dec tree impl
This an attempt to optimize the naive implementation.
Results are not particularly great, the naive implementation remains
very fast. The prim implementation seems to perform slightly better
on server class CPUs and consumes 3 times less mem.
The binbin one is definetely not working well, it shows better perf
only on very large trees.