Add feature logging
Use the search_ext to tell the plugin which model/featureset to collect.
Feature values are extracted as part of the search response using a
special field _ltrlog.
Multiple models/sets can be collected at once.
Feature collection happens during the fetch phase by running the feature
queries on the hits returned.

Queries to log are identified by using the _name attribute when used inside
the main search query or by an index in the list of rescore queries

For users that can afford running the feature queries a second time on the
returned hits this should permit to do runtime logging.