Remove usage of native script
this was a mean to share context built in the plugin with our query builders.
They decided to deprecte it so use another method to do this.

Drawbacks: it requires using WrappedQueryBuilder in tests to force the use
of XContent builders that will have the plugin context injected to it.