Update featureset with features provided in request body
Allows to update featureset by adding features in the request body
rather than providing a prefix query.
Initially it was only possible to update the featureset by adding
features from the store.
It's now possible to provide the feature definitions directly in
the request body bypassing the feature store.
Added an option to merge the features, it permits to replace a
feature instead of throwing an error.

relates #83