Add support for feature validation
Introduce a new json block to ask the plugin to fully validate features
prior to storing elements.
This still requires the user to explicitly set the "validation" which
contains the information needed to validate features:
- the index to run the query on (for mapping validation in some cases)
- the template param values

NOTE: that I had to change the body format to include the new "validation".
See changes in the docs folder for more details.

It might be interesting (1.1?) to add a new specific rest endpoint to validate
stored features/set/models.

closes #84