Fix build with 6.1.1
Use less/greater or equal in StoredLtrQueryIT (we only wnat to check score sign here)
Restored the prefix sub field as it is needed for large feature stores with
more than 1024 features.
Embbeded the missing token filters inside the plugin itself, this is not ideal
but this is just glue code to existing lucene filters.
Incremented the mapping version to 2 to be able to identify old stores.
There's no need to upgrade existing stores as the tokens indexed are similar.
Increased the default replica count from 0-1 to 0-2, 1 seemed a bit low.