Workaround removal of Scorer#tf()
Introduce a set of Query/Weight/Scorer suited to explore the postings.
This is still heavely relying on Weight#extractTerms() and would ideally
require a bigger refactoring to regain control of the user query string.
Ideally we should accept a query string not a random elastic query.
The reason is that it messes up the lucene workflow since we need to
be at the createWeight step while we should still be in the query building
phase.